176950434 move parsing of the license configs into tokenstat.py
