+++ src/licensemanager2/agent/settings.py
@@ -3,15 +3,12 @@ Configuration of the agent running in the cluster
 """
 
 from enum import Enum
-from functools import lru_cache
 from pathlib import Path
 import sys
-import typing
 
 from pkg_resources import get_supported_platform, resource_filename
-from pydantic import BaseSettings, DirectoryPath, Field, validator
+from pydantic import BaseSettings, DirectoryPath, Field
 from pydantic.error_wrappers import ValidationError
-from pydantic.main import BaseModel
 
 from licensemanager2.agent import logger
 
@@ -30,58 +27,11 @@ class LogLevelEnum(str, Enum):
 
 _JWT_REGEX = r"[a-zA-Z0-9+/]+\.[a-zA-Z0-9+/]+\.[a-zA-Z0-9+/]"
 _URL_REGEX = r"http[s]?://.+"
+_ADDR_REGEX = r"\S+?:\S+?:\d+"
+_SERVICE_ADDRS_REGEX = rf"{_ADDR_REGEX}(\s+{_ADDR_REGEX})*"
 _DEFAULT_BIN_PATH = Path(
     resource_filename("licensemanager2.agent", get_supported_platform())
 )
-_ADDR_REGEX = r"\S+?:\S+?:\d+"
-_SERVICE_ADDRS_REGEX = rf"{_ADDR_REGEX}(\s+{_ADDR_REGEX})*"
-
-
-class LicenseService(BaseModel):
-    """
-    A license service such as "flexlm", with a set of host-port tuples
-    representing the network location where the service is listening.
-    """
-
-    name: str
-    hostports: typing.List[typing.Tuple[str, int]]
-
-
-class LicenseServiceCollection(BaseModel):
-    """
-    A collection of LicenseServices, mapped by their names
-    """
-
-    services: typing.Dict[str, LicenseService]
-
-    @classmethod
-    def from_env_string(cls, services):
-        """
-        @returns LicenseServiceCollection from parsing colon-separated env input
-
-        The syntax is:
-
-        - servicename:host:port e.g. "flexlm:172.0.1.2:2345"
-
-        - each entry separated by spaces e.g.
-          "flexlm:172.0.1.2:2345 abclm:172.0.1.3:2319"
-
-        - if the same service appears twice in the list they will be
-          merged, e.g.:
-          "flexlm:173.0.1.2:2345 flexlm:172.0.1.3:2345"
-          -> (pseudodata) "flexlm": [("173.0.1.2", 2345), "173.0.1.3", 2345)]
-        """
-        self = cls(services={})
-        services = services.split()
-        for item in services:
-            name, host, port = item.split(":", 2)
-
-            svc = self.services.setdefault(
-                name, LicenseService(name=name, hostports=[])
-            )
-            svc.hostports.append((host, int(port)))
-
-        return self
 
 
 class _Settings(BaseSettings):
@@ -118,20 +68,8 @@ class _Settings(BaseSettings):
     class Config:
         env_prefix = "LM2_AGENT_"
 
-    @lru_cache
-    def _service_addrs(self) -> LicenseServiceCollection:
-        """
-        Convert the string form into a LicenseServiceCollection
-        """
-        return LicenseServiceCollection.from_env_string(self.SERVICE_ADDRS)
 
-    @property
-    def service_addrs(self) -> LicenseServiceCollection:
-        return self._service_addrs()
-
-
-@lru_cache
-def init_settings():
+def init_settings() -> _Settings:
     try:
         return _Settings()
     except ValidationError as e:
